Moved all coredata_test to device_test
